Rename document-signin-next-field to document-signing-next-field
